refactoring modals
